+++ modules/App/assets/vue-components/field-wysiwyg.js
@@ -16,7 +16,12 @@ export default {
         info: 'Rich text field',
         icon: 'system:assets/icons/wysiwyg.svg',
         render(value, field, context) {
-            return App.utils.truncate(App.utils.stripTags(value), context == 'table-cell' ? 20 : 50);
+
+            if (Array.isArray(value)) {
+                return value.length ? `${value.length}x...` : '';
+            }
+
+            return value ? App.utils.truncate(App.utils.stripTags(value), context == 'table-cell' ? 20 : 50) : '';
         }
     },