+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## WIP
 
 - Fix tree model nesting items
+- Fix table preview for repeatable wysiwyg fields
 
 ## 2.3.5 (2022-12-15)